:new: refactoring, voting plugin
