+++ config_files/exp_config.json
@@ -1,7 +1,8 @@
 {
-  "name": "database",
+  "name": "local_test",
   "host": "0.0.0.0",
   "port": 5010,
   "debug": "True",
-  "reset_db": "False"
+  "reset_db": "True",
+  "modules": ["news", "voting"]
 }
